+++ jsolex/src/main/resources/whats-new.md
@@ -6,6 +6,7 @@
 - Added geometry corrected image (no contrast enhancement) to the quick mode
 - Improve performance of SER conversion
 - Let user use decimal numbers for focal length and slew rates
+- Add sunscan to the list of setups
 
 ## What's New in Version 2.11.1